requirements.txt
absl-py==0.15.0
alembic==1.8.1
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
contextlib2==21.6.0
cryptography==39.0.2
cycler==0.11.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
flatbuffers==1.12
flit_core==3.6.0
fonttools==4.33.2
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.3.6
MarkupSafe==2.1.1
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pip==22.3.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyotp==2.7.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
setuptools==65.6.3
SimpleITK==2.1.1.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.64.0
typing_extensions==4.4.0
urllib3==1.26.9
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.0
